chore: add natspec comments to some functions

+++ packages/contracts/contracts/Interfaces/IDefaultPool.sol
@@ -4,13 +4,15 @@ pragma solidity 0.6.11;
 
 import "./IPool.sol";
 
-
 interface IDefaultPool is IPool {
     // --- Events ---
     event TroveManagerAddressChanged(address _newTroveManagerAddress);
-    event DefaultPoolZUSDDebtUpdated(uint _ZUSDDebt);
-    event DefaultPoolETHBalanceUpdated(uint _ETH);
+    event DefaultPoolZUSDDebtUpdated(uint256 _ZUSDDebt);
+    event DefaultPoolETHBalanceUpdated(uint256 _ETH);
 
     // --- Functions ---
-    function sendETHToActivePool(uint _amount) external;
+
+    /// @notice Send ETH to Active Pool
+    /// @param _amount ETH to send
+    function sendETHToActivePool(uint256 _amount) external;
 }
